==> notebooks/01_data_preprocess.ipynb <==
# Notebook 01 - Data Preprocess

### Scrape La Liga data from FBRef

import sys
sys.path.append('../src')

from data_preprocess import scrape_fbref_la_liga_data
from paths import RAW_DATA_DIR

# Parameters
start_year = 2018
end_year = 2023
output_path = RAW_DATA_DIR / 'data_gathered.csv'

# Scraping data
scrape_fbref_la_liga_data(start_year, end_year, output_path)

#### Visualize scraped data

import pandas as pd

df = pd.read_csv('/Users/ricardoheredia/Desktop/mle-soccer-project/data/raw/data_gathered.csv')
df

### Data Cleaning

import numpy as np
import pandas as pd
from janitor import clean_names
df = clean_names(df)
df

##### Manual correction

df.rename(columns = {'wk':'week',
                     'xg':'xG',
                     'xg_1':'xG_1'}, inplace=True)
df

##### General vision

df.info()

##### Data types

df.head(5)

df.dtypes

###### Correction

df['week'] = df['week'].astype('Int64')
df['date'] = pd.to_datetime(df['date'])

df.dtypes

#### Unique values

df.nunique().sort_values()

#### Dupes

df.duplicated().sum()

#### Separate categorical and numeric values

##### Categorical

cat = df.select_dtypes(exclude = 'number').copy()

cat.head()

##### Numerical

num = df.select_dtypes(include='number').copy()

num.head()

#### Manage categorical values

##### Null values

cat.isna().sum().sort_values(ascending = False)

cat.drop(columns = 'time',inplace = True)
cat.head(5)

#### Manage numerical values

##### Null values

num.isna().sum().sort_values(ascending = False)

num[num['attendance'].isnull()]

num.drop(columns = 'attendance',inplace = True)
num.head(5)

##### Outliers

num_std = 4

def outliers_std_dev(variable, num_std_dev=4):
    """
    Function to identify outliers based on standard deviation.

    Parameters:
    - variable: The variable containing numerical data (Pandas Series).
    - num_std_dev: The number of standard deviations to use as the threshold (int, default=4).

    Returns:
    - The indices of the values that are considered outliers (list).
    """
    # Remove null values for now
    variable = variable.dropna()
    # Calculate the limits
    mean = np.mean(variable)
    std_dev = np.std(variable)
    threshold = std_dev * num_std_dev
    lower_limit = mean - threshold
    upper_limit = mean + threshold
    # Find the indices of the values outside the limits
    indices = [index for index, value in variable.items() if value < lower_limit or value > upper_limit]
    return indices

def outlier_count(df, variable, num_std_dev=4):
    outliers = outliers_std_dev(df[variable], num_std_dev)
    return df.loc[outliers, variable].value_counts()


var_outliers_std = ['xG','xG_1']

for variable in var_outliers_std :
    print('\n' + variable + ':\n')
    print(outlier_count(num,variable,num_std))

## Save datasets after data quality

data_quality = pd.concat([num, cat], axis=1)

data_quality = data_quality[['week', 'date', 'day', 'home', 'score', 'away', 'xG', 'xG_1','venue', 'referee']]
data_quality

from paths import TRANSFORMED_DATA_DIR

output_path_data_quality = TRANSFORMED_DATA_DIR / 'data_quality.csv'

# Guardar el DataFrame como un archivo CSV
data_quality.to_csv(output_path_data_quality, index=False)

print(f"Data saved to {output_path_data_quality}")
