==> notebooks/03_feature_engineering.ipynb <==
## Notebook 03 - Feature Engineering

# Load libraries
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
%matplotlib inline
from sklearn.preprocessing import OneHotEncoder
from sklearn.preprocessing import OrdinalEncoder
from sklearn.preprocessing import LabelEncoder
from category_encoders import TargetEncoder

import sys
sys.path.append('../src')
from paths import TRANSFORMED_DATA_DIR

#Automcomplete
%config IPCompleter.greedy=True

# Load the cleaned dataset
data_quality = pd.read_csv(TRANSFORMED_DATA_DIR / 'data_quality.csv')

data_quality

### Create features and the target variable

The objective of the training is to predict match results, either Home wins or Away wins, so in this section we'll be creating the features and target variables needed to build and train the model prediction dataset.

An important note: I decided to train the model on the seasons before de 2023/2024 and validate it with this last season, since my initial idea is to predict the matches for the upcoming La Liga 2024/2025 season, though since there's no official match calendar and, at the time I'm doing this project, the Segunda División Playoffs are not done, so we don't know which will be the teams that are going to ascend.

### Feature engineering and data transformation

# Split the 'Score' column into 'HomeGoals' and 'AwayGoals' columns
data_quality[['home_goals', 'away_goals']] = data_quality['score'].str.split('–', expand=True).astype(float)

data_quality

# Create a season_start column so we can later train the 
data_quality['date'] = pd.to_datetime(data_quality['date'])
data_quality['season_start'] = data_quality['date'].apply(lambda x: x.year - 1 if x.month < 8 else x.year)

data_quality

# Create a target variable for the match result
def determine_result(row):
    if row['home_goals'] > row['away_goals']:
        return 'Home win'
    elif row['home_goals'] < row['away_goals']:
        return 'Away win'
    else:
        return 'Draw'

data_quality['result'] = data_quality.apply(determine_result, axis=1)

data_quality

#### One Hot Encoding

# Encode 'day' as a categorical feature using pd.get_dummies
data_quality['day'] = data_quality['date'].dt.day_name()
data_quality = pd.get_dummies(data_quality, columns=['day'])

# Display the updated dataset with new features
data_quality.head()

### Creating new features

# Let's create some new features
data_quality.reset_index(drop=True, inplace=True)
data_quality.sort_values(['date'], inplace=True)

data_quality

### Rolling averages

Since we're predicting match outcomes, this is a feature that'll help us achieve that. These are averages calculated over a moving window of data points, helping to smooth out short-term variations and reveal longer-term trends.

# First, let's isolate teams matches, so we can then calculate the rolling averages properly
for x in data_quality.home.unique():
    temp_df = data_quality[(data_quality['home'] == x) | (data_quality['away'] == x)]
    break

temp_df

# Update the rolling average goals for the home team in the main dataset (data_quality)
for x in data_quality.home.unique():
    temp_df = data_quality[(data_quality['home'] == x) | (data_quality['away'] == x)]
    temp_df = temp_df.sort_values(['date'])
    
    temp_df['goal_value_to_calculate'] = temp_df.apply(lambda y: y['home_goals'] if y['home'] == x else y['away_goals'], axis=1)
    temp_df['rolling_avg_goals'] = temp_df['goal_value_to_calculate'].rolling(window=5, closed="left", min_periods=1).mean()
    
    for index, row in temp_df.iterrows():
        if row['home'] == x:
            data_quality.at[index, 'home_rolling_avg_goals'] = row['rolling_avg_goals']
        else:
            data_quality.at[index, 'away_rolling_avg_goals'] = row['rolling_avg_goals']

data_quality[(data_quality['home'] == 'Real Madrid') | (data_quality['away'] == 'Real Madrid')][['week', 'date', 'home', 'away', 'home_goals', 'away_goals','home_rolling_avg_goals', 'away_rolling_avg_goals']]

# Do the same thing for xG
for x in data_quality.home.unique():
    temp_df = data_quality[(data_quality['home'] == x) | (data_quality['away'] == x)]
    temp_df = temp_df.sort_values(['date'])
    
    temp_df['xG_value_to_calculate'] = temp_df.apply(lambda y: y['xG'] if y['home'] == x else y['xG_1'], axis=1)
    temp_df['rolling_avg_xG'] = temp_df['xG_value_to_calculate'].rolling(window=5, closed="left", min_periods=1).mean()
    
    for index, row in temp_df.iterrows():
        if row['home'] == x:
            data_quality.at[index, 'home_rolling_avg_xG'] = row['rolling_avg_xG']
        else:
            data_quality.at[index, 'away_rolling_avg_xG'] = row['rolling_avg_xG']

# One last bit of clean up is to drop the rows where the rolling averages are null
data_quality = data_quality.dropna(subset=['home_rolling_avg_goals', 'away_rolling_avg_goals', 'home_rolling_avg_xG', 'away_rolling_avg_xG'])

data_quality[(data_quality['home'] == 'Real Madrid') | (data_quality['away'] == 'Real Madrid')][['date', 'home', 'away', 'xG', 'xG_1','home_rolling_avg_xG', 'away_rolling_avg_xG']]

pd.set_option('display.max_columns', None)
data_transformed = data_quality
data_transformed

from paths import TRANSFORMED_DATA_DIR

output_path_data_transformed = TRANSFORMED_DATA_DIR / 'data_transformed.csv'

# Guardar el DataFrame como un archivo CSV
data_transformed.to_csv(output_path_data_transformed, index=False)

print(f"Data saved to {output_path_data_transformed}")
