==> notebooks/06_inference.ipynb <==
# Notebook 06 - Inference

import warnings
warnings.filterwarnings("ignore")

import sys
sys.path.append('../src')

import os
import hopsworks
from comet_ml import API, APIExperiment, Model
import pandas as pd
import xgboost as xgb
from dotenv import load_dotenv
from paths import PARENT_DIR

# Load API keys from .env
load_dotenv(PARENT_DIR / '.env')
HOPSWORKS_API_KEY = os.environ['HOPSWORKS_API_KEY']
HOPSWORKS_PROJECT_NAME = 'rheredia8'
COMET_API_KEY = os.environ['COMET_API_KEY']
COMET_PROJECT_NAME = 'mle-soccer-project'
COMET_WORKSPACE = 'rheredia8'
COMET_MODEL_NAME = 'prime_church_4035'

# Hopsworks configs
project = hopsworks.login(api_key_value=HOPSWORKS_API_KEY)
feature_store = project.get_feature_store()

# Obtener el grupo de características
FEATURE_GROUP_NAME = 'la_liga_features'
FEATURE_GROUP_VERSION = 1
feature_group = feature_store.get_feature_group(
    name=FEATURE_GROUP_NAME,
    version=FEATURE_GROUP_VERSION
)

# Cargar los datos transformados
data_transformed = feature_group.read()
data_transformed.head()
